b2brocker/requirements.txt
asgiref==3.8.1
coverage==7.6.2
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
mimesis==18.0.0
mysqlclient==2.2.4
sqlparse==0.5.1
